Initial setup and changes to Twittler
Created .gitignore, moved script from HTML to application.js for cleanliness. Within the HTML I created a div to display the tweets and a button to fetch and display new tweets. Within the JS file, I moved the loop to display tweets to its own function so it can easily be executed by clicking the button.